Pain in the neck detail on Contextmenu/Delete in a table that also allows editing:
The act of invoking this action via a context menu
may also start a cell editor on the PV name.
The 'refresh' caused by the item removal
will cause that editor to think that somebody just entered
a name.
The result: It re-enters the PV Name that we just deleted!

Now the delete action cancels any open editor.