o.c.startup: Application code had a side effect in the 'module' extension point lookup. When it was for example looking for "startupParameters", it would in fact instantiate all classes that extend the o.c.s.module extension point, then check if the created class instance has the correct type. This means that it would actually load plugins, which in turn might read preferences, set the workspace (*), ... only to then notice that it was the wrong type. Now it only instantiates classes that were contributed for the requested extension point element.
*: Setting the workspace was a problem for the SNS product.
For example, the extension for o.c.startup.module.workspace might be just fine, but while looking for its implementation, the class for the .. workbench implementation was loaded, only to then be ignored because obvisouly didn't match the type for the workbench ext. point.
Still, having loaded the workbench class once meant activating its plugin and dependencies, which loaded preferences, which set the workspace, and then the workspace extension point could no longer adjust the workspace.