o.c.utility.pv.epics: Clarification on AccessRightsListener
The MonitorListener is already called when access rights change, but in
tests the write permissions don't always seem to be up to date in there.
The additional AccessRightsListener, on the other hand, always sees the
correct write permissions. So while it's often just an 'extra' value
update, it seems necessary.