o.c.security: Support initial 'system' login even with LDAP via prefs
When using a JAAS config file, that file is supposed to include entries
'unix' and 'windows' to allow the initial automated login based on the
OS, using the current user. When configuring the JAAS settings based on
Eclipse preferences, the 'unix' and 'windows' entries were not
available, so the automated system login failed. Now 'unix' and
'windows' entries are included when using the Eclipse preferences to set
JAAS.