o.c.alarm.beast: Fix update of newly added PVs
Code tried to suppress PV updates when config changed in case the alarm
state appeared the same. This fails if a GUI adds a new PV, and it's not
in error: _Other_ GUIs would actually show the new PV, but the alarm
tree of the GUI that added the PV would not show it until it actually
changes its alarm state.