o.c.opibuilder.editor/tools: Only one ImageIcons.decorator
Decorator for image icons which displays thumbnail of the image in the
icon was found in both the o.c.opibuilder.editor and ..tools plugins.
Kept the one in o.c.opibuilder.editor (tools depends on editor, would
not have reduced the number of dependencies to put it in tools).
While at it, added basic description to the implementation, fetch the
path 'ext' only once, and log exception to log.